fix(root): tailwind configurado pra funcionar no App.tsx e arquivos modificados para ficarem um nível acima, removendo a pasta design-system
