--- design-system/src/lib/styles/index.scss
@@ -1,9 +1,0 @@
-@import './variables';
-@import './typography';
-@import './reset.scss';
-
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-@import "@radix-ui/themes/styles.css";
