+++ src/App.tsx
@@ -1,0 +1,9 @@
+function App() {
+  return (
+    <>
+      <button className="flex bg-indigo-900 text-white p-4">This test</button>
+    </>
+  )
+}
+
+export default App
