+++ src/index.css
@@ -1,0 +1,2 @@
+@import "tailwindcss";
+@import './lib/styles/index.scss';
