+++ src/lib/styles/_variables.scss
