+++ src/lib/styles/index.scss
@@ -1,0 +1,5 @@
+@use './variables';
+@use './typography';
+@use './reset.scss';
+
+@import "@radix-ui/themes/styles.css";
