+++ src/stories/assets/accessibility.png
