+++ src/stories/assets/addon-library.png
