+++ src/stories/assets/assets.png
