+++ src/stories/assets/avif-test-image.avif
