+++ src/stories/assets/figma-plugin.png
