+++ src/stories/assets/styling.png
