+++ src/stories/assets/theming.png
