+++ src/vite-env.d.ts
